// offlinePWA/src/app/models/catalogo.items.enum.ts
export enum CatalogoItems {   
    componentes = "componentes",
    discapacidades = "discapacidades",
    estatutosMigratorios = "estatutosMigratorios",
    etnias = "etnias",
    generos = "generos",
    lgbtis = "lgbtis",
    motivosReferencias = "motivosReferencias",
    motivosSalidaPais = "motivosSalidaPais",
    nacionalidades = "nacionalidades",
    necesidadesEspecificasProteccion = "necesidadesEspecificasProteccion",
    necesidadesEspecificasHabitabilidad = "necesidadesEspecificasHabitabilidad",
    nivelesEscolaridad = "nivelesEscolaridad",
    nivelesParentesco = "nivelesParentesco",
    oficios = "oficios",
    organizacionesApoyo = "organizacionesApoyo",
    organizacionesReferencia = "organizacionesReferencia",
    profesiones = "profesiones",
    sexos = "sexos",
    situacionesMigratorias = "situacionesMigratorias",
    tiposAsistencia = "tiposAsistencia",
    tiposAyuda = "tiposAyuda",
    tiposDocumento = "tiposDocumento",
    tiposSeguimiento = "tiposSeguimiento",
    paises = "paises",
    ciudades = "ciudades",
    provincias = "provincias",
}